Vklücenije "-Wconversion".

+++ Compiler/src/lexical_analyzer.cpp
@@ -388,8 +388,8 @@ LexicalAnalysisResult LexicalAnalysis( const ProgramString& program_text )
 		}
 
 	push_lexem:
-		lexem.file_pos.line= line;
-		lexem.file_pos.pos_in_line= (unsigned int)( it - last_newline_it );
+		lexem.file_pos.line= static_cast<unsigned short>(line);
+		lexem.file_pos.pos_in_line= static_cast<unsigned short>( it - last_newline_it );
 
 		result.lexems.emplace_back( std::move(lexem) );
 
@@ -398,8 +398,8 @@ LexicalAnalysisResult LexicalAnalysis( const ProgramString& program_text )
 	Lexem eof_lexem;
 	eof_lexem.type= Lexem::Type::EndOfFile;
 	eof_lexem.text= "EOF"_SpC;
-	eof_lexem.file_pos.line= line;
-	eof_lexem.file_pos.pos_in_line= (unsigned int)( it - last_newline_it );
+	eof_lexem.file_pos.line= static_cast<unsigned short>(line);
+	eof_lexem.file_pos.pos_in_line= static_cast<unsigned short>( it - last_newline_it );
 
 	result.lexems.emplace_back( std::move(eof_lexem) );
